third_party/blink/web_tests/platform/win10/external/wpt/css/css-fonts/cjk-kerning-expected.txt

This is a testharness.js-based test.
PASS expected match: .kernON .latin vs .paltOFFkernON .latin
PASS expected match: .kernON .latin vs .paltONkernON .latin
PASS expected match: .kernOFF .latin vs .paltONkernOFF .latin
FAIL expected match: .kernON .cjk vs .paltONkernON .cjk assert_equals: expected 316 but got 384
PASS expected match: .default .cjk vs .kernOFF .cjk
PASS expected mismatch: .kernOFF .latin vs .kernON .latin
FAIL expected mismatch: .kernOFF .cjk vs .kernON .cjk assert_not_equals: got disallowed value 384
PASS expected mismatch: .paltOFFkernON .cjk vs .paltONkernON .cjk
PASS .default .latin matches one of [.kernON .latin, .kernOFF .latin]
Harness: the test ran to completion.
